Move purview/utils/teiid.py to scrapers/teiid.py
